test(revive): keep the Revive cases beside session.go
relaunch_test.go keeps the pane probe helper it covers.
