+++ internal/sessioncmd/session_test.go
@@ -847,3 +847,45 @@ func TestDeleteGroupTakesItsSubtreeAndKeepsNesting(t *testing.T) {
 		t.Fatalf("moving the subtree unhooked the terminal from its agent: parent %q", after.ParentID)
 	}
 }
+
+// An agent that exits leaves its window open on the shell it was launched
+// from. Revive puts the tool back inside that pane instead of refusing the
+// row as still running, which is what the manager's own revive key does.
+func TestReviveRestartsTheAgentInsideItsLivePane(t *testing.T) {
+	h := newSessionHarness(t)
+	created, err := h.sessions.Create(h.caller.ID, CreateSessionOptions{Name: "worker", Prompt: "hold the line"})
+	if err != nil {
+		t.Fatalf("Create: %v", err)
+	}
+	waitForSessionOutput(t, h.sessions, h.caller.ID, created.ID, "hold the line")
+	waitForAgentGone(t, h.driver, created.ID)
+
+	revived, err := h.sessions.Revive(h.caller.ID, created.ID)
+	if err != nil {
+		t.Fatalf("Revive: %v", err)
+	}
+	if !revived.Running || revived.Status != status.Starting {
+		t.Fatalf("revived session = %+v", revived)
+	}
+	screen := waitForSessionOutput(t, h.sessions, h.caller.ID, created.ID, "resumed")
+	if !strings.Contains(screen.Output, "hold the line") {
+		t.Fatalf("an in-pane revive keeps what the pane already held, got %q", screen.Output)
+	}
+	if !h.driver.Exists(created.ID) {
+		t.Fatal("revive should have left the window running")
+	}
+}
+
+func TestReviveRefusesWhileTheAgentIsStillRunning(t *testing.T) {
+	h := newSessionHarness(t)
+	created, err := h.sessions.Create(h.caller.ID, CreateSessionOptions{Name: "chatty", Tool: "resting"})
+	if err != nil {
+		t.Fatalf("Create: %v", err)
+	}
+	waitForSessionOutput(t, h.sessions, h.caller.ID, created.ID, "❯")
+
+	if _, err := h.sessions.Revive(h.caller.ID, created.ID); err == nil ||
+		!strings.Contains(err.Error(), "still running") {
+		t.Fatalf("reviving a session whose agent is up = %v", err)
+	}
+}
